+++ src/test/kotlin/in/kkkev/jjidea/vcs/changes/ChangeIdRevisionNumberTest.kt
@@ -1,8 +1,6 @@
 package `in`.kkkev.jjidea.vcs.changes
 
 import `in`.kkkev.jjidea.jj.ChangeId
-import io.kotest.matchers.comparables.shouldBeGreaterThan
-import io.kotest.matchers.comparables.shouldBeLessThan
 import io.kotest.matchers.shouldBe
 import org.junit.jupiter.api.Test
 
@@ -26,12 +24,12 @@ class ChangeIdRevisionNumberTest {
     }
 
     @Test
-    fun `different revisions compare by string`() {
+    fun `different revisions compare as non-zero`() {
         val rev1 = ChangeIdRevisionNumber(ChangeId("vvv", "v"))
         val rev2 = ChangeIdRevisionNumber(ChangeId("www", "w"))
 
-        rev1 shouldBeLessThan rev2
-        rev2 shouldBeGreaterThan rev1
+        rev1.compareTo(rev2) shouldBe -1
+        rev2.compareTo(rev1) shouldBe -1
     }
 
     companion object {